split actions code
