+++ ReoGrid/Actions/BackupRangeInfo.cs
@@ -1,0 +1,31 @@
+﻿/*****************************************************************************
+ * 
+ * ReoGrid - Opensource .NET Spreadsheet Control
+ * 
+ * https://reogrid.net/
+ *
+ * THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY
+ * KIND, EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR
+ * PURPOSE.
+ *
+ * Thank you to all contributors!
+ * 
+ * (c) 2012-2020 Jingwood, unvell.com <jingwood at unvell.com>
+ * 
+ ****************************************************************************/
+
+namespace unvell.ReoGrid.Actions
+{
+	internal struct BackupRangeInfo
+	{
+		internal int start;
+		internal int count;
+
+		public BackupRangeInfo(int start, int count)
+		{
+			this.start = start;
+			this.count = count;
+		}
+	}
+}
