+++ ReoGrid/ReoGrid.csproj
@@ -242,10 +242,45 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Actions\AddOutlineAction.cs" />
+    <Compile Include="Actions\AutoFillSerialAction.cs" />
+    <Compile Include="Actions\BaseOutlineAction.cs" />
+    <Compile Include="Actions\BaseWorksheetAction.cs" />
+    <Compile Include="Actions\CollapseOutlineAction.cs" />
+    <Compile Include="Actions\CopyRangeAction.cs" />
+    <Compile Include="Actions\CreateAutoFilterAction.cs" />
+    <Compile Include="Actions\ExpandOutlineAction.cs" />
+    <Compile Include="Actions\HideColumnsAction.cs" />
+    <Compile Include="Actions\HideRowsAction.cs" />
+    <Compile Include="Actions\InsertWorksheetAction.cs" />
+    <Compile Include="Actions\MergeRangeAction.cs" />
+    <Compile Include="Actions\MoveRangeAction.cs" />
+    <Compile Include="Actions\OutlineAction.cs" />
+    <Compile Include="Actions\RemoveColumnsAction.cs" />
+    <Compile Include="Actions\RemoveOutlineAction.cs" />
+    <Compile Include="Actions\RemoveRangeBorderAction.cs" />
+    <Compile Include="Actions\RemoveRangeDataAction.cs" />
+    <Compile Include="Actions\RemoveWorksheetAction.cs" />
+    <Compile Include="Actions\SetCellDataAction.cs" />
+    <Compile Include="Actions\SetColumnsWidthAction.cs" />
+    <Compile Include="Actions\SetPartialGridAction.cs" />
+    <Compile Include="Actions\SetRangeBorderAction.cs" />
+    <Compile Include="Actions\SetRangeDataAction.cs" />
+    <Compile Include="Actions\SetRangeDataFormatAction.cs" />
+    <Compile Include="Actions\SetRangeStyleAction.cs" />
+    <Compile Include="Actions\SetRowsHeightAction.cs" />
+    <Compile Include="Actions\UnhideColumnsAction.cs" />
+    <Compile Include="Actions\UnhideRowsAction.cs" />
+    <Compile Include="Actions\UnmergeRangeAction.cs" />
+    <Compile Include="Actions\WorkbookAction.cs" />
+    <Compile Include="Actions\WorksheetActionGroup.cs" />
+    <Compile Include="Actions\WorksheetReusableAction.cs" />
+    <Compile Include="Actions\WorksheetReusableActionGroup.cs" />
     <Compile Include="Android\Graphics.cs" />
     <Compile Include="Android\Renderer.cs" />
     <Compile Include="Android\ReoGridView.cs" />
     <Compile Include="Android\Platform.cs" />
+    <Compile Include="Actions\BackupRangeInfo.cs" />
     <Compile Include="CellTypes\ButtonCell.cs" />
     <Compile Include="CellTypes\CellBody.cs" />
     <Compile Include="CellTypes\CellTypesManager.cs" />
@@ -336,6 +371,8 @@
     <Compile Include="Graphics\Point.cs" />
     <Compile Include="Graphics\Rectangle.cs" />
     <Compile Include="Graphics\Size.cs" />
+    <Compile Include="Actions\InsertRowsAction.cs" />
+    <Compile Include="Actions\InsertColumnsAction.cs" />
     <Compile Include="Interaction\KeyCode.cs" />
     <Compile Include="Interaction\MouseButtons.cs" />
     <Compile Include="Interaction\CursorStyle.cs" />
@@ -359,6 +396,9 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <Compile Include="Actions\RemoveRangeStyleAction.cs" />
+    <Compile Include="Actions\StepRangeFontSizeAction.cs" />
+    <Compile Include="Actions\RemoveRowsAction.cs" />
     <Compile Include="Test\RichTextTestForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -376,7 +416,7 @@
     <Compile Include="WinForm\AboutForm.Designer.cs">
       <DependentUpon>AboutForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Actions.cs" />
+    <Compile Include="Actions\ClearOutlineAction.cs" />
     <Compile Include="Chart\Interface.cs" />
     <Compile Include="Common\ActionManager.cs" />
     <Compile Include="Common\GraphicsToolkit.cs" />
